Move `get_eviction_dispatcher` to worker_intents.py - thanks @radix
