+++ otter/worker_intents.py
@@ -2,9 +2,11 @@
 Intents for worker code including the controller, supervisor, and launch server
 worker.
 """
+from functools import partial
 
 from characteristic import attributes
 
+from effect import TypeDispatcher
 from effect.twisted import deferred_performer
 
 
@@ -25,3 +27,15 @@ def perform_evict_server(supervisor, dispatcher, intent):
         intent.log, intent.transaction_id,
         intent.scaling_group.tenant_id,
         intent.server_id)
+
+
+def get_eviction_dispatcher(supervisor):
+    """
+    Get a dispatcher with :class:`EvictServerFromScalingGroup`'s performer.
+
+    :param reactor: Twisted reactor
+    :param supervisor: a :class:`otter.supervisor.ISupervisor` provider
+    """
+    return TypeDispatcher({
+        EvictServerFromScalingGroup: partial(perform_evict_server, supervisor)
+    })
